Split SwingConsolePane#main to separate class
The logic is going to get more involved shortly...
