+++ src/test/java/org/scijava/ui/swing/console/SwingConsolePaneBenchmark.java
@@ -1,0 +1,67 @@
+/*
+ * #%L
+ * SciJava UI components for Java Swing.
+ * %%
+ * Copyright (C) 2010 - 2015 Board of Regents of the University of
+ * Wisconsin-Madison.
+ * %%
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * 
+ * 1. Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ * 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+ * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDERS OR CONTRIBUTORS BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+ * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
+ * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
+ * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ * #L%
+ */
+
+package org.scijava.ui.swing.console;
+
+import org.scijava.Context;
+import org.scijava.thread.ThreadService;
+import org.scijava.ui.UIService;
+
+/**
+ * A behavioral test and benchmark of {@link SwingConsolePane}.
+ *
+ * @author Curtis Rueden
+ */
+public class SwingConsolePaneBenchmark {
+
+	// -- Main method --
+
+	/** A manual test drive of the Swing UI's console pane. */
+	public static void main(final String[] args) throws Exception {
+		final Context context = new Context();
+		context.service(UIService.class).showUI();
+
+		System.out.println("Hello!");
+
+		final ThreadService threadService = context.service(ThreadService.class);
+		threadService.run(new Runnable() {
+
+			@Override
+			public void run() {
+				System.out.println("This is a test of the emergency console system.");
+				System.err.println("In a real emergency, your computer would explode.");
+			}
+
+		}).get();
+
+		System.err.println("Goodbye!");
+	}
+
+}
